backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
greenlet==1.1.3
idna==3.3
numpy==1.23.2
pandas==1.4.3 # utilizado no generator_service.py
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40 # utilizado no generator_service.py
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
watchdog==2.1.9 # utilizado no generator_service.py
